requirements.txt
clip==0.2.0
numpy==1.24.3
Pillow==10.4.0
PyYAML==6.0.2
scipy==1.14.1
torch==1.10.0+cu113
torchvision==0.11.0+cu113
tqdm==4.64.1
